// src/utils/enums/validation.enum.ts
export enum ValidationEnum {
  NotString = 'Должно быть строкой',
  NotInt = 'Должно быть числом',
  NotUUID = 'Не соответствует типу UUID',
  NotMin = 'Должно превышать единицу',
  NotMax = 'Превышает максимум в 100 единиц товара',
  TooBigCommentLength = 'Комментарий не должен быть больше, чем 255 символов',
  NotEmail = 'Введён некорректный почтовый адрес',
  IncorrectPhoneNumber = 'Введён некорректный номер телефона',
  NotArray = 'Должно быть массивом',
  NotPromoType = 'Не является типом акции',
  NotCallbackStatus = 'Не является статусом заявки на обратный звонок',
  NotOrderStatus = 'Не является статусом заказа',
  NotRoleType = 'Не является ролью пользователя',
  NotJWT = 'Не является JWT',
  NotFile = 'Не является файлом',
  NameNotEmpty = 'Название не должно быть пустым',
  NotValueType = 'Не соответствует ожидаемому типу данных',
}